feat: create current menu id custom hook
