+++ apps/web/src/services/advanced/AdvancedLSB.vue
@@ -1,22 +1,15 @@
 <script setup lang="ts">
 import { computed, reactive } from 'vue';
-import { useRoute } from 'vue-router/composables';
-
-import { clone } from 'lodash';
 
 import { MENU_ID } from '@/lib/menu/config.js';
+
+import { useCurrentMenuId } from '@/common/composables/current-menu-id';
 
 import BookmarkLSB from '@/services/advanced/components/BookmarkLSB.vue';
 import DomainSettingsLSB from '@/services/advanced/components/PreferencesLSB.vue';
 
-const route = useRoute();
-
 const state = reactive({
-    menuId: computed<string>(() => {
-        const reversedMatched = clone(route.matched).reverse();
-        const closestRoute = reversedMatched.find((d) => d.meta?.menuId !== undefined);
-        return closestRoute?.meta?.menuId;
-    }),
+    menuId: computed<string>(() => useCurrentMenuId()),
 });
 </script>
