+++ apps/web/src/services/asset-inventory/AssetInventoryLSB.vue
@@ -1,23 +1,16 @@
 <script setup lang="ts">
 import { computed, reactive } from 'vue';
-import { useRoute } from 'vue-router/composables';
-
-import { clone } from 'lodash';
 
 import { MENU_ID } from '@/lib/menu/config.js';
+
+import { useCurrentMenuId } from '@/common/composables/current-menu-id';
 
 import CloudServiceLSB from '@/services/asset-inventory/components/CloudServiceLSB.vue';
 import MetricExplorerLSB from '@/services/asset-inventory/components/MetricExplorerLSB.vue';
 import SecurityLSB from '@/services/asset-inventory/components/SecurityLSB.vue';
 
-const route = useRoute();
-
 const state = reactive({
-    menuId: computed<string>(() => {
-        const reversedMatched = clone(route.matched).reverse();
-        const closestRoute = reversedMatched.find((d) => d.meta?.menuId !== undefined);
-        return closestRoute?.meta?.menuId;
-    }),
+    menuId: computed<string>(() => useCurrentMenuId()),
 });
 </script>
