+++ apps/web/src/services/cost-explorer/CostExplorerLSB.vue
@@ -2,10 +2,9 @@
 import { computed, reactive } from 'vue';
 import { useRoute } from 'vue-router/composables';
 
-import { clone } from 'lodash';
-
-import type { MenuId } from '@/lib/menu/config';
 import { MENU_ID } from '@/lib/menu/config';
+
+import { useCurrentMenuId } from '@/common/composables/current-menu-id';
 
 import CostAdvancedSettingsLSB from '@/services/cost-explorer/components/CostAdvancedSettingsLSB.vue';
 import CostAnalysisLSB from '@/services/cost-explorer/components/CostAnalysisLSB.vue';
@@ -13,12 +12,7 @@ import CostAnalysisLSB from '@/services/cost-explorer/components/CostAnalysisLSB
 const route = useRoute();
 
 const state = reactive({
-    selectedMenuId: computed(() => {
-        const reversedMatched = clone(route.matched).reverse();
-        const closestRoute = reversedMatched.find((d) => d.meta?.menuId !== undefined);
-        const targetMenuId: MenuId = closestRoute?.meta?.menuId || MENU_ID.WORKSPACE_HOME;
-        return targetMenuId;
-    }),
+    selectedMenuId: computed<string>(() => useCurrentMenuId()),
 
 });
 </script>
